style: Refactor Tooltip class, #8
BREAKING CHANGE: `contentClassName` has been renamed `containerClassName` to be more consistent
